Add files via upload
The room thermostat is a device used to control the temperature in a 
room or camera, keeping it at a value set by the user. These devices are 
mainly used to control heating and cooling systems such as HVAC systems 
(heating, ventilation and air conditioning), to ensure adequate thermal comfort in a space 
interior. The project is implemented in the Proteus program (where I made the electronic diagram) and in the Keil uVision program (where I made the operating code, both in C code and in assembly code).